move printSetupNotes
