+++ packages/cli/src/commands/setup/deploy/helpers/index.js
@@ -1,7 +1,6 @@
 import fs from 'fs'
 import path from 'path'
 
-import boxen from 'boxen'
 import execa from 'execa'
 import { Listr } from 'listr2'
 
@@ -136,19 +135,6 @@ export const addToDotEnvTask = ({ lines }) => {
       }
 
       fs.appendFileSync(env, lines.join('\n'))
-    },
-  }
-}
-
-export const printSetupNotes = (notes) => {
-  return {
-    title: 'One more thing...',
-    task: (_ctx, task) => {
-      task.title = `One more thing...\n\n ${boxen(notes.join('\n'), {
-        padding: { top: 1, bottom: 1, right: 1, left: 1 },
-        margin: 1,
-        borderColour: 'gray',
-      })}  \n`
     },
   }
 }
