+++ packages/cli/src/commands/setup/deploy/providers/baremetal.js
@@ -5,9 +5,9 @@ import { Listr } from 'listr2'
 
 import { errorTelemetry } from '@redwoodjs/telemetry'
 
-import { addPackagesTask, getPaths } from '../../../../lib'
+import { addPackagesTask, getPaths, printSetupNotes } from '../../../../lib'
 import c from '../../../../lib/colors'
-import { addFilesTask, printSetupNotes } from '../helpers'
+import { addFilesTask } from '../helpers'
 import { DEPLOY, ECOSYSTEM, MAINTENANCE } from '../templates/baremetal'
 
 export const command = 'baremetal'
