+++ packages/cli/src/commands/setup/ui/libraries/windicss.js
@@ -6,9 +6,8 @@ import { Listr } from 'listr2'
 
 import { errorTelemetry } from '@redwoodjs/telemetry'
 
-import { getPaths } from '../../../../lib'
+import { getPaths, printSetupNotes } from '../../../../lib'
 import c from '../../../../lib/colors'
-import { printSetupNotes } from '../helpers'
 
 export const command = 'windicss'
 export const aliases = ['windi']
